added dummy data in users db, products db, vendors db, reviews db and implemented single product api

+++ src/main/java/com/panic/sasserver/model/Payment.java
@@ -17,12 +17,12 @@ public class Payment{
         Long orderId;
         Long userId;
         Long vendorId;
-        Integer amount;
+        Double amount;
         LocalDate paymentDate;
         @Enumerated
         PaymentStatus status;
 
-        public Payment(PaymentStatus status, LocalDate paymentDate, Integer amount, Long vendorId, Long userId, Long orderId) {
+        public Payment(PaymentStatus status, LocalDate paymentDate, Double amount, Long vendorId, Long userId, Long orderId) {
                 this.status = status;
                 this.paymentDate = paymentDate;
                 this.amount = amount;
